requirements.txt
langgraph
langchain
openai
faiss-cpu
python-dotenv
langchain-community
